+++ Gemfile
@@ -4,7 +4,7 @@ gem 'rails', '4.2.5.2'
 
 # NB: this *must* be by Git ref; else will break asset versioning in
 #     config/initializers/assets.rb, preventing app startup
-gem 'europeana-styleguide', github: 'europeana/europeana-styleguide-ruby', ref: 'a386cb1'
+gem 'europeana-styleguide', github: 'europeana/europeana-styleguide-ruby', ref: 'ab30cfa'
 
 # Use a forked version of stache with downstream changes, until merged upstream
 # @see https://github.com/agoragames/stache/pulls/rwd
@@ -33,10 +33,8 @@ gem 'paper_trail', '~> 4.0'
 gem 'puma', '~> 2.13'
 gem 'rack-rewrite'
 gem 'rails_admin', '~> 0.8.0'
-gem 'rails_admin_globalize_field', '~> 0.4'
 gem 'redis-rails', '~> 4.0'
 gem 'redis-rails-instrumentation'
-gem 'responders', '~> 2.1'
 gem 'sass-rails'
 gem 'soundcloud', '~> 0.3'
 gem 'therubyracer'
