+++ app/assets/stylesheets/rails_admin/custom/theming.scss
@@ -1,0 +1,3 @@
+.navbar-nav > li > form {
+  padding: 15px;
+}
